# tests/snapshots/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Data Platform Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/repo/diff": {
      "get": {
        "operationId": "repo_diff_api_v1_repo_diff_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repo Diff",
        "tags": [
          "Repo"
        ]
      }
    },
    "/api/v1/repo/refs": {
      "get": {
        "operationId": "repo_refs_api_v1_repo_refs_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "upstream_ref",
            "required": false,
            "schema": {
              "default": "upstream/main",
              "title": "Upstream Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repo Refs",
        "tags": [
          "Repo"
        ]
      }
    },
...
    "/api/v1/repo/status": {
      "get": {
        "operationId": "repo_status_api_v1_repo_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repo Status",
        "tags": [
          "Repo"
        ]
      }
    },
    "/api/v1/repo/upstream/connect": {
      "post": {
        "description": "Attach upstream remote + persist config (per workspace job).",
        "operationId": "upstream_connect_api_v1_repo_upstream_connect_post",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "repo_url",
            "required": true,
            "schema": {
              "title": "Repo Url",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "branch",
            "required": false,
            "schema": {
              "default": "main",
              "title": "Branch",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upstream Connect",
        "tags": [
          "Repo"
        ]
      }
    },
    "/api/v1/repo/upstream/diff": {
      "get": {
        "description": "direction:\n  - upstream_to_workspace: git diff upstream_ref..HEAD\n  - workspace_to_upstream: git diff HEAD..upstream_ref\npaths:\n  - comma-separated scope prefixes, e.g. jobs,dags,configs\n  - .copilot is always excluded",
        "operationId": "upstream_diff_api_v1_repo_upstream_diff_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated path prefixes to include",
            "in": "query",
            "name": "paths",
            "required": false,
            "schema": {
              "default": "jobs,dags,configs",
              "description": "Comma-separated path prefixes to include",
              "title": "Paths",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "upstream_to_workspace",
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_chars",
            "required": false,
            "schema": {
              "default": 20000,
              "maximum": 500000,
              "minimum": 1000,
              "title": "Max Chars",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upstream Diff",
        "tags": [
          "Repo"
        ]
      }
    },
    "/api/v1/repo/upstream/fetch": {
      "post": {
        "description": "Fetch upstream refs based on saved upstream.json.",
        "operationId": "upstream_fetch_api_v1_repo_upstream_fetch_post",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upstream Fetch",
        "tags": [
          "Repo"
        ]
      }
    },
    "/api/v1/repo/upstream/status": {
      "get": {
        "description": "Show ahead/behind vs upstream branch + file drift summary.",
        "operationId": "upstream_status_api_v1_repo_upstream_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated path prefixes to include (e.g., app,ui,copilot_spec.yaml)",
            "in": "query",
            "name": "paths",
            "required": false,
            "schema": {
              "default": "jobs,dags,configs",
              "description": "Comma-separated path prefixes to include (e.g., app,ui,copilot_spec.yaml)",
              "title": "Paths",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upstream Status",
        "tags": [
          "Repo"
        ]
      }
    },
    "/api/v1/repo/upstream/switch": {
      "post": {
        "description": "Upstream switching modes:\n  - replace: update upstream remote/branch (baseline unchanged)\n  - rebaseline: replace + set baseline to current HEAD\n  - reset: replace + hard reset workspace to upstream ref + baseline = new HEAD (destructive)",
        "operationId": "upstream_switch_api_v1_repo_upstream_switch_post",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "description": "Optional if .copilot/upstream.json already exists",
            "in": "query",
            "name": "repo_url",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional if .copilot/upstream.json already exists",
              "title": "Repo Url"
            }
          },
          {
            "description": "Optional if .copilot/upstream.json already exists",
            "in": "query",
            "name": "branch",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional if .copilot/upstream.json already exists",
              "title": "Branch"
            }
          },
          {
            "description": "Optional if .copilot/upstream.json already exists",
            "in": "query",
            "name": "remote",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional if .copilot/upstream.json already exists",
              "title": "Remote"
            }
          },
          {
            "description": "replace|rebaseline|reset",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "replace",
              "description": "replace|rebaseline|reset",
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "description": "Only used for reset mode to allow dirty repos",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only used for reset mode to allow dirty repos",
              "title": "Force",
              "type": "boolean"
            }
          },
          {
            "description": "Only used for reset mode; runs git clean -fd after reset",
            "in": "query",
            "name": "clean_untracked",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only used for reset mode; runs git clean -fd after reset",
              "title": "Clean Untracked",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upstream Switch",
        "tags": [
          "Repo"
        ]
      }
    },
...
    "/api/v2/repo/diff": {
      "get": {
        "operationId": "repo_diff_api_v2_repo_diff_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repo Diff",
        "tags": [
          "Repo"
        ]
      }
    },
    "/api/v2/repo/refs": {
      "get": {
        "operationId": "repo_refs_api_v2_repo_refs_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "upstream_ref",
            "required": false,
            "schema": {
              "default": "upstream/main",
              "title": "Upstream Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repo Refs",
        "tags": [
          "Repo"
        ]
      }
    },
...
    "/api/v2/repo/status": {
      "get": {
        "operationId": "repo_status_api_v2_repo_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repo Status",
        "tags": [
          "Repo"
        ]
      }
    },
    "/api/v2/repo/upstream/connect": {
      "post": {
        "description": "Attach upstream remote + persist config (per workspace job).",
        "operationId": "upstream_connect_api_v2_repo_upstream_connect_post",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "repo_url",
            "required": true,
            "schema": {
              "title": "Repo Url",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "branch",
            "required": false,
            "schema": {
              "default": "main",
              "title": "Branch",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upstream Connect",
        "tags": [
          "Repo"
        ]
      }
    },
...
  }
}